third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/content-pseudo-element-dynamic-class-expected.txt

CONSOLE ERROR: line 14: Uncaught TypeError: host1.createShadowRoot is not a function
